% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphDS1.R
\name{coxphDS1}
\alias{coxphDS1}
\title{Cox proportional-hazards model.}
\usage{
coxphDS1(survival_time, survival_event, terms, method, data)
}
\arguments{
\item{survival_time}{survivial time}

\item{survival_event}{survivial event}

\item{terms}{terms for the model}

\item{method}{method for tie handling}

\item{data}{a character, the name of an optional data frame containing the variables in 
in the \code{formula}.}
}
\value{
the coxph model.
}
\description{
Cox proportional-hazards model.
}
\author{
Inberg, G.
}
