% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphDS3.R
\name{coxphDS3}
\alias{coxphDS3}
\title{Cox proportional-hazards model.}
\usage{
coxphDS3(survival_time, survival_event, terms, beta.vect, data)
}
\arguments{
\item{survival_time}{survivial time}

\item{survival_event}{survivial event}

\item{terms}{terms for the model}

\item{beta.vect}{model weight parameters}

\item{data}{a character, the name of an optional data frame containing the variables in 
in the \code{formula}.}
}
\value{
a list of aggregated statistics based on local data.
}
\description{
Cox proportional-hazards model.
}
\author{
Inberg, G.
}
