\name{gee.alphaphi.ds}
\alias{gee.alphaphi.ds}
\title{GEE}
\usage{
  gee.alphaphi.ds(formula, family, link, data, cluster,
    corstr, start.betas = NULL)
}
\arguments{
  \item{formula}{}

  \item{family}{}

  \item{link}{}

  \item{data}{}

  \item{cluster}{}

  \item{corstr}{}

  \item{start.betas}{}
}
\description{
  GEE helper function
}

