\name{gee.beta.ds}
\alias{gee.beta.ds}
\title{GEE helper 2}
\usage{
  gee.beta.ds(formula, family, link, data, cluster, alpha,
    phi, corstr, start.betas = NULL, zcor = NULL)
}
\arguments{
  \item{formula}{}

  \item{family}{}

  \item{link}{}

  \item{data}{}

  \item{cluster}{}

  \item{alpha}{}

  \item{phi}{}

  \item{corstr}{}

  \item{start.betas}{}

  \item{zcor}{}
}
\description{
  GEE helper function
}

