% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmDS2.R
\name{glmDS2}
\alias{glmDS2}
\title{Fits a generalized linear model}
\usage{
glmDS2(formula, family, beta.vect, offset, weights, data)
}
\arguments{
\item{formula}{an object of class \code{formula}.}

\item{family}{a character, description of the error distribution and link function to used in the model.}

\item{beta.vect}{a string character: the starting values for the parameters in the linear predictor.}

\item{offset}{a character, null or a numeric vector that can be used to specify an a priori known component 
to be included in the linear predictor during fitting.}

\item{weights}{a character, the name of an optional vector of 'prior weights' to be used in the fitting 
process. Should be NULL or a numeric vector.}

\item{data}{a character, the name of an optional data frame containing the variables in 
in the \code{formula}.}
}
\value{
a list which contains: the fitted \code{family}, a score vector and an information matrix
}
\description{
Fits a GLM for one iteration only
}
\details{
This function cannot be called through 'datashield.aggregate' like other aggregate functions
It is only called from within the client function \code{ds.glm} in the package \code{dsModellingClient}
}
\author{
Gaye, A.; Burton, P.; Laflamme, P.
}
