% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lexisDS2.R
\name{lexisDS2}
\alias{lexisDS2}
\title{TODO: document lexisDS2}
\usage{
lexisDS2(datatext = NULL, intervalWidth, maxmaxtime, idCol, entryCol, exitCol,
  statusCol, vartext = NULL)
}
\description{
TODO: document lexisDS2
}
\details{

}

