% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphDS1.R
\name{coxphDS1}
\alias{coxphDS1}
\title{Distributed Cox model learning, local initialization.}
\usage{
coxphDS1(data, survival_time, terms)
}
\arguments{
\item{data}{a character, the name of the data frame that holds the data.}

\item{survival_time}{character, the survivial time variable}

\item{terms}{comma separated string containing the terms for the model}
}
\value{
a list of time values and aggregated statistics based on local data.
}
\description{
Distributed Cox model learning, local initialization.
}
\author{
Inberg, G.
}
