% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphDS2.R
\name{coxphDS2}
\alias{coxphDS2}
\title{Distributed Cox model learning, local initialization based on global unique time values.}
\usage{
coxphDS2(data, survival_time, survival_event, terms, tuniq)
}
\arguments{
\item{data}{a character, the name of the data frame that holds the data.}

\item{survival_time}{character, the survivial time variable}

\item{survival_event}{character, the survivial event variable}

\item{terms}{comma separated string containing the terms for the model}

\item{tuniq}{comma separated string containing the unique time values}
}
\value{
a list of aggregated statistics based on local data.
}
\description{
Distributed Cox model learning, local initialization based on global unique time values.
}
\author{
Inberg, G.
}
