% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphDS2.R
\name{coxphDS2}
\alias{coxphDS2}
\title{Distributed Cox model learning, local initialization based on global unique time values.}
\usage{
coxphDS2(survival_time, survival_event, terms, tuniq, data)
}
\arguments{
\item{survival_time}{survivial time}

\item{survival_event}{survivial event}

\item{terms}{terms for the model}

\item{tuniq}{unique time values}

\item{data}{a character, the name of an optional data frame containing the variables in 
in the \code{formula}.}
}
\value{
a list of aggregated statistics based on local data.
}
\description{
Distributed Cox model learning, local initialization based on global unique time values.
}
\author{
Inberg, G.
}
