% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphDS3.R
\name{coxphDS3}
\alias{coxphDS3}
\title{Distributed Cox model learning step3, calculate aggregated statistics..}
\usage{
coxphDS3(data, survival_time, terms, beta.vect)
}
\arguments{
\item{data}{a character, the name of the data frame that holds the data.}

\item{survival_time}{character, the survivial time variable}

\item{terms}{comma separated string containing the terms for the model}

\item{beta.vect}{comma separated string containing the model weight parameters}
}
\value{
a list of aggregated statistics based on local data.
}
\description{
Distributed Cox model learning step3, calculate aggregated statistics..
}
\author{
Inberg, G.
}
