\name{glm.ds}
\alias{glm.ds}
\title{Fits a generalized linear model}
\usage{
  glm.ds(formula, family, beta.vect = NULL)
}
\arguments{
  \item{formula}{an object of class \code{formula}}

  \item{family}{a description of the error distribution and
  link function to used in the model}

  \item{beta.vect}{starting values for the parameters in
  the linear predictor}
}
\description{
  Generalized linear modellling
}

